Moved code to satisfy Sonar limits
